arch: Hal abstraction for TTY + privelege managers moved into arch

+++ kernel/src/core/memory/heap.c
@@ -6,7 +6,7 @@
 #define HEAP_SIZE_CLASSES_COUNT 13
 
 static struct mutex heap_mutex;
-static uint32_t heap_size_classes[HEAP_SIZE_CLASSES_COUNT] = {
+static size_t heap_size_classes[HEAP_SIZE_CLASSES_COUNT] = {
 	16, 32, 64, 128, 256, 512, 1024, 2048, 4096, 8192, 16384, 32768, 65536};
 
 struct heap_slub_obj_hdr {
@@ -15,7 +15,7 @@ struct heap_slub_obj_hdr {
 
 static struct heap_slub_obj_hdr *slubs[HEAP_SIZE_CLASSES_COUNT];
 
-static uint32_t heap_get_size_class(uint32_t size) {
+static size_t heap_get_size_class(size_t size) {
 	for (size_t i = 0; i < HEAP_SIZE_CLASSES_COUNT; ++i) {
 		if (size <= heap_size_classes[i]) {
 			return i;
@@ -24,16 +24,16 @@ static uint32_t heap_get_size_class(uint32_t size) {
 	return HEAP_SIZE_CLASSES_COUNT;
 }
 
-static bool heap_add_objects_to_slubs(uint32_t index) {
-	uint32_t size = heap_size_classes[index];
-	uint32_t objects_count = BLOCK_SIZE / size;
-	uint32_t block = hal_phys_krnl_alloc_area(BLOCK_SIZE);
+static bool heap_add_objects_to_slubs(size_t index) {
+	size_t size = heap_size_classes[index];
+	size_t objects_count = BLOCK_SIZE / size;
+	size_t block = hal_phys_krnl_alloc_area(BLOCK_SIZE);
 	if (block == 0) {
 		return false;
 	}
 	for (size_t i = 0; i < objects_count; ++i) {
 		size_t offset = i * size;
-		uint32_t address = HAL_VIRT_KERNEL_MAPPING_BASE + block + offset;
+		size_t address = HAL_VIRT_KERNEL_MAPPING_BASE + block + offset;
 		struct heap_slub_obj_hdr *block = (struct heap_slub_obj_hdr *)address;
 		block->next = slubs[index];
 		slubs[index] = block;
@@ -48,14 +48,14 @@ void heap_init() {
 	}
 }
 
-void *heap_alloc(uint32_t size) {
+void *heap_alloc(size_t size) {
 	if (size == 0) {
 		return NULL;
 	}
 	mutex_lock(&heap_mutex);
-	uint32_t size_class = heap_get_size_class(size);
+	size_t size_class = heap_get_size_class(size);
 	if (size_class == HEAP_SIZE_CLASSES_COUNT) {
-		uint32_t result =
+		uintptr_t result =
 			hal_phys_krnl_alloc_area(ALIGN_UP(size, HAL_VIRT_PAGE_SIZE));
 		if (result == 0) {
 			mutex_unlock(&heap_mutex);
@@ -77,15 +77,15 @@ void *heap_alloc(uint32_t size) {
 	return result;
 }
 
-void heap_free(void *area, uint32_t size) {
+void heap_free(void *area, size_t size) {
 	if (area == NULL) {
 		return;
 	}
 	mutex_lock(&heap_mutex);
-	uint32_t size_class = heap_get_size_class(size);
+	size_t size_class = heap_get_size_class(size);
 	if (size_class == HEAP_SIZE_CLASSES_COUNT) {
 		mutex_unlock(&heap_mutex);
-		hal_phys_krnl_free_area((uint32_t)area,
+		hal_phys_krnl_free_area((uintptr_t)area,
 								ALIGN_UP(size, HAL_VIRT_PAGE_SIZE));
 		return;
 	}
