+++ kernel/src/core/memory/heap.h
@@ -4,8 +4,8 @@
 #include <utils/utils.h>
 
 void heap_init();
-void *heap_alloc(uint32_t size);
-void heap_free(void *area, uint32_t size);
+void *heap_alloc(size_t size);
+void heap_free(void *area, size_t size);
 
 #define ALLOC_OBJ(t) (t *)heap_alloc(sizeof(t))
 #define FREE_OBJ(p) heap_free(p, sizeof(typeof(*(p))))
